[REF] account: Split big files account.py & account_view.xml
