+++ addons/account/__manifest__.py
@@ -32,7 +32,18 @@ You could use this simplified accounting in case you work with an (external) acc
         'wizard/setup_wizards_view.xml',
         'wizard/pos_box.xml',
         'views/partner_view.xml',
-        'views/account_view.xml',
+        'views/account_account_type_views.xml',
+        'views/account_account_views.xml',
+        'views/account_group_views.xml',
+        'views/account_journal_views.xml',
+        'views/account_account_tag_views.xml',
+        'views/account_bank_statement_views.xml',
+        'views/account_reconcile_model_views.xml',
+        'views/account_tax_views.xml',
+        'views/account_full_reconcile_views.xml',
+        'views/account_payment_term_views.xml',
+        'views/account_chart_template_views.xml',
+        'views/res_partner_bank_views.xml',
         'views/report_statement.xml',
         'views/account_report.xml',
         'data/mail_template_data.xml',
