+++ addons/account/models/__init__.py
@@ -2,7 +2,12 @@
 
 from . import sequence_mixin
 from . import partner
-from . import account
+from . import res_partner_bank
+from . import account_account_tag
+from . import account_account
+from . import account_journal
+from . import account_tax
+from . import account_tax_report
 from . import account_reconcile_model
 from . import account_payment_term
 from . import account_move
